Create array 3
3. For a given array, write functions to perform the following:
i. Finding single element in an array
ii. Finding multiple elements in an array
iii. Finding duplicates in a sorted array
iv. Finding duplicates in an unsorted array
v. Finding a pair of elements with sum k
vi. Finding a pair of elements with sum k in sorted array
vii. Finding max and min in a single scan